# src/controllers/v3/openapi.yaml
components:
...
    schemas:
...
        ActivityDetail:
            properties:
                id:
                    type: string
                message:
                    type: string
                params:
                    properties: {}
                    additionalProperties:
                        type: string
                    type: object
                activityAt: {}
                blob: {}
            required:
                - message
                - params
                - activityAt
            type: object
            additionalProperties: false
...
        GroupPermission:
            properties:
                resource:
                    type: string
                scopes:
                    items:
                        type: string
                    type: array
            required:
                - scopes
            type: object
            additionalProperties: false
        GroupRole:
            properties:
                name:
                    type: string
                permissions:
                    items:
                        $ref: '#/components/schemas/GroupPermission'
                    type: array
            required:
                - name
                - permissions
            type: object
            additionalProperties: false
        GroupAccess:
            properties:
                name:
                    type: string
                parent:
                    type: string
                roles:
                    items:
                        $ref: '#/components/schemas/GroupRole'
                    type: array
            required:
                - roles
            type: object
            additionalProperties: false
        UserReference:
            properties:
                id:
                    type: string
                email:
                    type: string
            type: object
            additionalProperties: false
        GroupMember:
            properties:
                member:
                    $ref: '#/components/schemas/UserReference'
                roles:
                    items:
                        type: string
                    type: array
            required:
                - member
                - roles
            type: object
            additionalProperties: false
        GroupMembership:
            properties:
                name:
                    type: string
                parent:
                    type: string
                members:
                    items:
                        $ref: '#/components/schemas/GroupMember'
                    type: array
            type: object
            additionalProperties: false
        OrgNamespace:
            properties:
                name:
                    type: string
                orgUnit:
                    type: string
                enabled:
                    type: boolean
                updatedAt:
                    type: number
                    format: double
            required:
                - name
                - orgUnit
                - enabled
                - updatedAt
            type: object
            additionalProperties: false
...
    securitySchemes:
        jwt:
            type: oauth2
            description: 'Authz Client Credential'
            flows:
                clientCredentials:
                    tokenUrl: 'https://token_endpoint'
                    scopes: {}
        portal:
            type: http
            description: 'Authz Portal Login'
            scheme: bearer
            bearerFormat: JWT
        openid:
            type: openIdConnect
            description: 'OIDC Login'
            openIdConnectUrl: 'https://well_known_endpoint'
info:
    title: 'APS Directory API'
    version: 3.0.0
    description: 'API Services Portal by BC Gov API Programme Services'
    license:
        name: MIT
    contact:
        name: 'BC Gov APS'
openapi: 3.0.0
paths:
...
    '/organizations/{org}':
        get:
            operationId: organization-units
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema: {}
            tags:
                - Organizations
            security: []
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
    '/organizations/{org}/roles':
        get:
            operationId: get-organization-roles
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GroupAccess'
            description: '> `Required Scope:` GroupAccess.Manage'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - GroupAccess.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
    '/organizations/{org}/access':
        get:
            operationId: get-organization-access
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GroupMembership'
            description: '> `Required Scope:` GroupAccess.Manage'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - GroupAccess.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
        put:
            operationId: put-organization-access
            responses:
                '204':
                    description: 'No content'
            description: '> `Required Scope:` GroupAccess.Manage'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - GroupAccess.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GroupMembership'
    '/organizations/{org}/gateways':
        get:
            operationId: organization-gateways
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/OrgNamespace'
                                type: array
            description: '> `Required Scope:` Namespace.Assign'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
    '/organizations/{org}/{orgUnit}/gateways/{gatewayId}':
        put:
            operationId: assign-namespace-to-organization
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    result: {type: string}
                                required:
                                    - result
                                type: object
            description: '> `Required Scope:` Namespace.Assign'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: orgUnit
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: enable
                    required: false
                    schema:
                        default: true
                        type: boolean
        delete:
            operationId: unassign-namespace-from-organization
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    result: {type: string}
                                required:
                                    - result
                                type: object
            description: '> `Required Scope:` Namespace.Assign'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: orgUnit
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
    '/organizations/{org}/activity':
        get:
            operationId: org-namespace-activity
            responses:
                '200':
                    description: 'Activity[]'
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/ActivityDetail'
                                type: array
            description: '> `Required Scope:` Namespace.Assign'
            summary: 'Get Namespace Activity for gateways associated with this Organization Unit'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: first
                    required: false
                    schema:
                        default: 20
                        format: double
                        type: number
                -
                    in: query
                    name: skip
                    required: false
                    schema:
                        default: 0
                        format: double
                        type: number
...
servers:
    -
        url: /ds/api/v3
...